extract parse_notes and parse_prerequisites into separate files
